REF/TST: directories for CategoricalIndex, RangeIndex tests (#30476)
